+++ DBAdmin/09_create-WhoIsActive-collector-proc.sql
@@ -18,10 +18,10 @@ BEGIN
 	*/
 	CREATE TABLE #ResultsTemp
 	([collection_time]       DATETIME NOT NULL,
-	 [session_id]            SMALLINT NOT NULL, 
+	 [session_id]            INT NOT NULL, 
 	 [status]                VARCHAR(30) NOT NULL, 
-	 [blocking_session_id]   SMALLINT NULL, 
-	 [blocked_session_count] SMALLINT NULL, 
+	 [blocking_session_id]   INT NULL, 
+	 [blocked_session_count] INT NULL, 
 	 [wait_info]             NVARCHAR(4000) NULL, 
 	 [database_name]         NVARCHAR(128) NULL, 
 	 [host_name]             NVARCHAR(128) NULL, 
@@ -39,7 +39,7 @@ BEGIN
 	 [query_plan]            XML NULL, 
 	 [used_memory]           BIGINT NOT NULL, 
 	 [tran_start_time]       DATETIME NULL, 
-	 [open_tran_count]       SMALLINT NULL, 
+	 [open_tran_count]       INT NULL, 
 	 [percent_complete]      REAL NULL, 
 	 [additional_info]       XML NULL, 
 	 [start_time]            DATETIME NOT NULL,
